Add data frame to Slides

Add data frame to Slides

Usage

# S3 method for class 'data.frame'
add_to_slides(
  object,
  presentation_id,
  on = NULL,
  object_id = new_id("table"),
  digits = NULL,
  overwrite = FALSE,
  ...
)

Arguments

object

A data.frame

presentation_id

The id from the Slides presentation

on

The id or number of the slide to add to

object_id

A unique id for the table

digits

the minimum number of significant digits, see format. If NULL getOption("digits") is used.

overwrite

If TRUE and an object with object_id exists it will deleted and replaced.

...

Not used in this method

Value

A presentation object after updating

Details

The data frame is added as a table with the column names in bold as the first row. For other formatting use the flextable package and add_to_slides.flextable.

Examples

if (FALSE) { # interactive()
# \donttest{
s <- choose_slides()
obj <- iris[1:5, ]
add_to_slides(obj, s, on = 1)
# }
}